refactor: promote ability loader upper
