+++ simaple/container/simulation.py
@@ -5,7 +5,7 @@ from typing import Callable, cast
 import pydantic
 
 from simaple.core import ExtendedStat, JobCategory, JobType
-from simaple.data.ability import get_best_ability
+from simaple.data import get_best_ability
 from simaple.data.baseline import get_baseline_gearset
 from simaple.data.builtin_strategy import get_builtin_strategy
 from simaple.data import get_damage_logic
