+++ tests/data/test_ability.py
@@ -1,5 +1,5 @@
 from simaple.core.jobtype import JobType
-from simaple.data.ability import get_best_ability
+from simaple.data import get_best_ability
 
 
 def test_get_ability():
